+++ Makefile
@@ -12,11 +12,11 @@ DB = lldb
 SOUND_LIBS = -lao -lmpg123
 CONF_LIBS = -lpthread #-lwiringPi
 
-$(BIN)/main: $(SRC)/main.c $(SRC)/play.c $(INC)/play.h
-	@$(CC) $(SRC)/main.c $(SRC)/play.c -o $(BIN)/main $(SOUND_LIBS) $(CONF_LIBS)
+$(BIN)/main: $(SRC)/main.c $(SRC)/play.c $(SRC)/read.c $(INC)/play.h $(INC)/read.h $(INC)/shared.h
+	@$(CC) $(SRC)/main.c $(SRC)/play.c $(SRC)/read.c -o $(BIN)/main $(SOUND_LIBS) $(CONF_LIBS)
 
-keys: $(SRC)/main.c $(SRC)/play.c $(INC)/play.h
-	@$(CC) $(SRC)/keys.c $(SRC)/play.c -o $(BIN)/main $(SOUND_LIBS) $(CONF_LIBS)
+keys: $(SRC)/main.c $(SRC)/play.c $(SRC)/read.c $(INC)/play.h $(INC)/read.h $(INC)/shared.h
+	@$(CC) $(SRC)/keys.c $(SRC)/play.c $(SRC)/read.c -o $(BIN)/main $(SOUND_LIBS) $(CONF_LIBS)
 
 run: $(BIN)/main
 	@$(BIN)/main
